Develop

+++ app/Http/Controllers/Frontend/CartController.php
@@ -18,6 +18,7 @@ use App\InvoiceItem;
 use App\Mail\InvoiceMail;
 use App\Promo;
 use App\Repositories\UserRepo;
+use App\Shipment;
 use App\UserComboCourse;
 use App\UserCourse;
 use Auth;
@@ -265,7 +266,7 @@ class CartController extends Controller
             $totalPrice += $item->amount * $item->price;
         }
 
-        if ($totalPrice > $user->money) {
+        if (!$request->input('shipment')['is_cod'] && $totalPrice > $user->money) {
             $request->validate([
                 'cart' => [
                     function ($attribute, $value, $fail) {
@@ -279,6 +280,7 @@ class CartController extends Controller
             DB::beginTransaction();
 
             $shipInfo = $request->input('ship_info');
+            $shipmentInfo = $request->input('shipment');
 
             // Process courses
             $coursesInCart = $cart->where('type', ObjectType::COURSE);
@@ -341,6 +343,17 @@ class CartController extends Controller
                     $invoiceItem->save();
                 }
 
+
+                if ($shipmentInfo && $shipInfo['shipping']) {
+                    $shipment = new Shipment();
+                    $shipment->partner = 'GHTK';
+                    $shipment->ship_money = $shipmentInfo['ship_money'];
+                    $shipment->is_fast = $shipmentInfo['is_fast'];
+                    $shipment->is_ship_cod = $shipmentInfo['is_cod'];
+                    $shipment->invoice_id = $invoice->id;
+                    $shipment->save();
+                }
+
                 $notificationInvoices[] = $invoice;
             }
 
@@ -390,7 +403,12 @@ class CartController extends Controller
 
             Cart::where('user_id', $user->id)->delete();
 
-            $this->userRepo->removeMoney($user->id, $totalPrice);
+            if ($shipmentInfo && $shipInfo['shipping']) {
+                $totalPrice += $shipmentInfo['ship_money'];
+            }
+            if ($shipmentInfo && !$shipmentInfo['is_cod']) {
+                $this->userRepo->removeMoney($user->id, $totalPrice);
+            }
 
             DB::commit();
